DFE-968 - created a new common service for getting publication basiuc details, broke out Data uplodas tab into separate component, and amended Manage Data responses to not include publicationTitle
