+++ src/explore-education-statistics-admin/src/pages/DummyReferenceData.ts
@@ -1,4 +1,4 @@
-import { IdLabelPair } from '@admin/services/common/types/types';
+import { IdLabelPair } from '@admin/services/common/types';
 
 export enum DateType {
   DayMonthYear,
