+++ src/explore-education-statistics-admin/src/pages/admin-dashboard/components/AdminDashboardPublicationsTab.tsx
@@ -3,7 +3,7 @@ import FormSelect from '@common/components/form/FormSelect';
 import React from 'react';
 import Accordion from '@common/components/Accordion';
 import AccordionSection from '@common/components/AccordionSection';
-import { IdLabelPair } from '@admin/services/common/types/types';
+import { IdLabelPair } from '@admin/services/common/types';
 import Link from '@admin/components/Link';
 import AdminDashboardPublicationSummary from './AdminDashboardPublicationSummary';
