+++ src/explore-education-statistics-admin/src/services/common/mock/axios-mock.ts
@@ -1,0 +1,14 @@
+import MockAdapter from 'axios-mock-adapter';
+
+export default async (mock: MockAdapter) => {
+  const mockData = (await import(
+    /* webpackChunkName: "mock-data" */ './mock-data'
+  )).default;
+
+  // getPublicationDetailsForRelease
+  mock.onGet(/\/release\/.*\/publication/).reply(({ url }) => {
+    const releaseIdMatch = url ? url.match(/\/release\/(.*)\/publication/) : [''];
+    const releaseId = releaseIdMatch ? releaseIdMatch[1] : '';
+    return [200, mockData.getPublicationDetailsForRelease(releaseId)];
+  });
+};
