+++ src/explore-education-statistics-admin/src/services/edit-release/data/types.ts
@@ -14,11 +14,6 @@ export interface DataFile {
   metadataFile: FileDetails;
 }
 
-export interface DataFileView {
-  publicationTitle: string;
-  dataFiles: DataFile[];
-}
-
 export interface UploadDataFilesRequest {
   subjectTitle: string;
   dataFile: File;
