+++ src/explore-education-statistics-admin/src/services/edit-release/setup/types.ts
@@ -1,7 +1,4 @@
-import {
-  DayMonthYearValues,
-  IdLabelPair,
-} from '@admin/services/common/types/types';
+import {DayMonthYearValues, IdLabelPair,} from '@admin/services/common/types';
 
 export interface ReleaseSetupDetails {
   id: string;
